+++ NEWS.md
@@ -1,3 +1,12 @@
+# Version (development version)
+
+## New Features
+
+ * The default value for R option `revdepcheck.extras.snapshot.source`
+   can now be set via environment variable
+   `R_REVDEPCHECK_EXTRAS_SNAPSHOT_SOURCE` on package load.
+   
+
 # Version 0.7.0 [2025-07-10]
 
 ## Significant changes